Avoid humongeous files

+++ src/track.py
@@ -15,29 +15,39 @@ import datetime
 MINDUR = 15 # Minumum duration of mix.
 MAXDUR = 15 * 60 # Maximum duration of mix.
 
-def generate():
-    targetduration = random.uniform(MINDUR, MAXDUR)
+def generate(mp3=True, play=False):
+    targetduration = int(random.uniform(MINDUR, MAXDUR))
 
     asset = random.choice([name for name in glob.glob("../_assets/*") if os.path.isdir(name)])
     asset = "../_assets/Axisem083_8" # TEST: Manual override.
-    print "Creating a track from '%s' asset, duration %s" % (asset, jumpcity.seconds2hhmmss(targetduration))
+    print "Creating a track from '%s' asset, target duration %s" % (asset, jumpcity.seconds2hhmmss(targetduration))
     fns = glob.glob(os.path.join(asset, "*"))
 
     mixercmd = "sox -m "
     for nr, fn in enumerate(random.sample(fns, 3)):
-        print "   Source: ", fn
-        # See how long the sample is, if it's longer than 'targetduration' seconds, fade it out
+        # See how long the sample is.
+        fndur = jumpcity.soundfileduration(fn)
+        print "    Source for layer%d: %s (duration %s)" % (nr, fn, jumpcity.seconds2hhmmss(fndur))
+        # If sample duration is longer than 'targetduration' seconds, fade it out.
         fade = ""
-        if jumpcity.soundfileduration(fn) > targetduration:
+        if fndur > targetduration:
             fade = "fade 0 %d 4" % targetduration
+            print "    ...fading to: %s" % fade
+        # If it is shorter than the target duration, repeat it.
+        repeat = ""
+        if fndur < targetduration:
+            repeats = (targetduration - fndur) / fndur
+            repeat = "repeat %d" % repeats
+            print "    ...repeating, extra %d times (for a total duration of %s)" % (repeats, (repeats + 1) * fndur)
         panning = (
             "remix 1 2", # No panning for sample #1
             "remix 1 2v0.25", # Right panning for sample #2
             "remix 1v0.25 2", # Left panning for sample #3
             )
         layerfn = "layer%d.wav" % nr
-        cmd = 'sox "%s" "%s" %s repeat 99 %s' % (fn, layerfn, panning[nr], fade)
+        cmd = 'sox "%s" "%s" %s %s %s' % (fn, layerfn, panning[nr], repeat, fade)
         os.system(cmd)    
+        print "    ...resulting duration: %s" % jumpcity.seconds2hhmmss(jumpcity.soundfileduration(layerfn))
         mixercmd += '"%s" ' % layerfn
 
     trackname = "track-%s.wav" % datetime.datetime.now().strftime("%Y-%m-%d.%H-%M-%S")
@@ -46,11 +56,15 @@ def generate():
     os.system(mixercmd)
     print "done"
 
-    os.system("play -q %s" % trackname)
+    os.system("rm layer*.wav") # Remove tempfiles.
 
-    print "Making mp3..."
-    os.system("lame %s" % trackname)
-    os.unlink(trackname)
+    if play:
+        os.system("play -q %s" % trackname)
+
+    if mp3:
+        print "Making mp3..."
+        os.system("lame %s" % trackname)
+        os.unlink(trackname)
 
 
 if __name__ == "__main__":
